requirements.txt
pandas>=2.2
openpyxl>=3.1
jinja2>=3.1
python-dotenv>=1.0
alpha-vantage>=2.3
